implement cabot_ui plugins
