+++ cabot_ui/cabot_ui/interface.py
@@ -28,8 +28,6 @@ from rclpy.exceptions import InvalidServiceNameException
 from rclpy.callback_groups import MutuallyExclusiveCallbackGroup
 from cabot_ui.cabot_rclpy_util import CaBotRclpyUtil
 
-from ament_index_python.packages import get_package_share_directory
-
 import std_msgs.msg
 import cabot_msgs.msg
 import cabot_msgs.srv
@@ -39,11 +37,6 @@ from cabot_ui.event import NavigationEvent
 from cabot_ui.turn_detector import Turn
 from cabot_ui.social_navigation import SNMessage
 from cabot_common import vibration
-
-from pydub import AudioSegment
-import pydub.playback
-
-import subprocess
 
 
 class SpeechPriority:
@@ -97,20 +90,6 @@ class UserInterface(object):
 
         self.last_social_announce = None
         self.last_notify_turn = {"directional_indicator": None, "vibrator": None}
-
-        self._enable_speaker = False
-        self._audio_dir = os.path.join(get_package_share_directory(os.environ.get("CABOT_SITE")), "sound")
-        self._audio_file_path = None
-        self._speaker_volume = 0.0
-        if os.path.isdir(self._audio_dir):
-            wav_files = [
-                f
-                for f in os.listdir(self._audio_dir)
-                if f.endswith(".wav") and os.path.isfile(os.path.join(self._audio_dir, f))
-            ]
-        else:
-            wav_files = []
-        self.audio_files = ",".join(wav_files)
 
     def _activity_log(self, category="", text="", memo="", visualize=False):
         log = cabot_msgs.msg.Log()
@@ -189,65 +168,6 @@ class UserInterface(object):
             CaBotRclpyUtil.info("speak requested")
         except InvalidServiceNameException as e:
             CaBotRclpyUtil.error(F"Service call failed: {e}")
-
-    def enable_speaker(self, enable_speaker: str):
-        self._enable_speaker = enable_speaker.lower() == "true"
-        self._activity_log("change speaker config", "enable speaker", str(self._enable_speaker), visualize=True)
-
-    def set_audio_file(self, filename):
-        self._audio_file_path = os.path.join(self._audio_dir, filename)
-        self._activity_log("change speaker config", "audio file", str(self._audio_file_path), visualize=True)
-
-    def set_speaker_volume(self, volume):
-        try:
-            speaker_volume = float(volume)
-        except Exception:
-            CaBotRclpyUtil.error(f"Invalid volume value: {volume}")
-            return
-
-        speaker_volume = max(0, min(100, speaker_volume))
-        self._speaker_volume = speaker_volume
-
-        try:
-            if speaker_volume == 0:
-                subprocess.run(
-                    ["pactl", "set-sink-mute", "@DEFAULT_SINK@", "on"],
-                    check=True
-                )
-            else:
-                subprocess.run(
-                    ["pactl", "set-sink-mute", "@DEFAULT_SINK@", "off"],
-                    check=True
-                )
-                subprocess.run(
-                    ["pactl", "set-sink-volume", "@DEFAULT_SINK@", f"{int(speaker_volume)}%"],
-                    check=True
-                )
-        except subprocess.CalledProcessError as e:
-            CaBotRclpyUtil.error(f"Failed to adjust volume via pactl: {e}")
-            return
-
-        self._activity_log("change speaker config", "volume (%)", str(self._speaker_volume), visualize=True)
-
-    def speaker_alert(self):
-        if not self._enable_speaker:
-            CaBotRclpyUtil.error("Speaker is disabled")
-            return
-
-        if not os.path.isfile(self._audio_file_path):
-            CaBotRclpyUtil.error(F"Audio file not found: {self._audio_file_path}")
-            return
-
-        if self._speaker_volume == 0:
-            CaBotRclpyUtil.info("Speaker is muted")
-            return
-
-        try:
-            sound = AudioSegment.from_wav(self._audio_file_path)
-            CaBotRclpyUtil.info(F"Playing {self._audio_file_path} (volume: {self._speaker_volume}%)")
-            pydub.playback.play(sound)
-        except Exception as e:
-            CaBotRclpyUtil.error(F"Playback failed: {e}")
 
     def vibrate(self, pattern=vibration.UNKNOWN):
         self._activity_log("cabot/interface", "vibration", vibration.get_name(pattern), visualize=True)
