+++ cabot_ui/scripts/cabot_ui_manager.py
@@ -34,6 +34,7 @@ Low-level (cabot_common.event) should be mapped into ui-level (cabot_ui.event)
 Author: Daisuke Sato<daisuke@cmu.edu>
 """
 
+from importlib.metadata import entry_points
 import signal
 import sys
 import threading
@@ -42,53 +43,74 @@ import yaml
 
 from rcl_interfaces.msg import ParameterDescriptor, ParameterType
 import rclpy
-import rclpy.client
-from rclpy.node import Node
-from rclpy.executors import SingleThreadedExecutor
-from rclpy.executors import MultiThreadedExecutor
 from rclpy.callback_groups import MutuallyExclusiveCallbackGroup
-from rclpy.qos import QoSProfile, QoSDurabilityPolicy
 import std_msgs.msg
 import std_srvs.srv
 
 import cabot_common.button
 from cabot_common.event import BaseEvent, ButtonEvent, ClickEvent, HoldDownEvent
+from cabot_ui.node_manager import NodeManager
 from cabot_ui.event import MenuEvent, NavigationEvent, ExplorationEvent
 from cabot_ui.menu import Menu
-from cabot_ui.status import State, StatusManager
 from cabot_ui.interface import UserInterface
-from cabot_ui.navigation import Navigation, NavigationInterface
+from cabot_ui.plugin import NavigationInterface
 from cabot_ui.cabot_rclpy_util import CaBotRclpyUtil
-from cabot_ui.description import Description
 
 from diagnostic_updater import Updater, FunctionDiagnosticTask
 from diagnostic_msgs.msg import DiagnosticStatus
 
 
+class NavigationPlugins():
+    def __init__(self, plugins, node_manager, delegate):
+        eps = entry_points(group="cabot_ui.plugins")
+        self._plugins = []
+
+        for name in plugins:
+            for ep in eps:
+                if ep.name != name:
+                    continue
+                CaBotRclpyUtil.info(f"Loading plugin {ep.name} {ep.value}")
+                cls = ep.load()
+                instance = cls(node_manager)
+                instance.delegate = delegate
+                self._plugins.append(instance)
+
+    @property
+    def plugins(self):
+        return self._plugins
+
+    @property
+    def ready(self):
+        for p in self._plugins:
+            if not p.ready:
+                return False
+        return True
+
+    def process_event(self, event):
+        for p in self._plugins:
+            if p.process_event(event):
+                break
+
+
 class CabotUIManager(NavigationInterface, object):
-    def __init__(self, node, nav_node, tf_node, srv_node, act_node, soc_node, desc_node):
-        self._node = node
+    def __init__(self, node_manager):
+        self._node = node_manager.get_node()
         self._logger = self._node.get_logger()
         CaBotRclpyUtil.initialize(self._node)
 
         self.in_navigation = False
-        self.destination = None
 
-        self.reset()
-
-        self._handle_button_mapping = node.declare_parameter('handle_button_mapping', 2).value
+        self._handle_button_mapping = self._node.declare_parameter('handle_button_mapping', 2).value
         if self._handle_button_mapping == 2:
             self._event_mapper = EventMapper2()
         else:
             self._event_mapper = EventMapper1()
         self._event_mapper.delegate = self
-        self._status_manager = StatusManager.get_instance()
-        self._status_manager.delegate = self
         self._interface = UserInterface(self._node)
         self._interface.delegate = self
-        self._navigation = Navigation(nav_node, tf_node, srv_node, act_node, soc_node)
-        self._navigation.delegate = self
-        self._description = Description(desc_node)
+
+        plugins = self._node.declare_parameter('navigation_plugins', ["navigation", "feature", "description", "speaker"]).value
+        self._navigation_plugins = NavigationPlugins(plugins, node_manager, self)
         # self._exploration = Exploration()
         # self._exploration.delegate = self
 
@@ -97,36 +119,10 @@ class CabotUIManager(NavigationInterface, object):
         self._node.create_subscription(std_msgs.msg.String, "/cabot/event", self._event_callback, 10, callback_group=MutuallyExclusiveCallbackGroup())
         self._eventPub = self._node.create_publisher(std_msgs.msg.String, "/cabot/event", 10, callback_group=MutuallyExclusiveCallbackGroup())
 
-        # request language
-        e = NavigationEvent("getlanguage", None)
-        msg = std_msgs.msg.String()
-        msg.data = str(e)
-        self._eventPub.publish(msg)
-
-        def handleside_callback(msg):
-            # request handleside
-            self.handleside = msg.data
-            self.send_handleside()
-        qos_profile = QoSProfile(depth=10)
-        qos_profile.durability = QoSDurabilityPolicy.TRANSIENT_LOCAL
-        self.handleside_sub = node.create_subscription(std_msgs.msg.String, "/cabot/features/handleside", handleside_callback, qos_profile)
-
-        def touchmode_callback(msg):
-            # request touchmode
-            self.touchmode = msg.data
-            self.send_touchmode()
-        qos_profile = QoSProfile(depth=10)
-        qos_profile.durability = QoSDurabilityPolicy.TRANSIENT_LOCAL
-        self.touchmode_sub = node.create_subscription(std_msgs.msg.String, "/cabot/features/touchmode", touchmode_callback, qos_profile)
-
-        self._touchModeProxy = self._node.create_client(std_srvs.srv.SetBool, "/cabot/set_touch_speed_active_mode", callback_group=MutuallyExclusiveCallbackGroup())
-
-        self._userSpeedEnabledProxy = self._node.create_client(std_srvs.srv.SetBool, "/cabot/user_speed_enabled", callback_group=MutuallyExclusiveCallbackGroup())
-
         self.updater = Updater(self._node)
 
         def manager_status(stat):
-            if self._navigation.i_am_ready:
+            if self._navigation_plugins.ready:
                 stat.summary(DiagnosticStatus.OK, "Ready")
             else:
                 stat.summary(DiagnosticStatus.ERROR, "Not ready")
@@ -135,33 +131,10 @@ class CabotUIManager(NavigationInterface, object):
 
         self.create_menu_timer = self._node.create_timer(1.0, self.create_menu, callback_group=MutuallyExclusiveCallbackGroup())
 
-        self.send_speaker_audio_files()
-
-    def send_handleside(self):
-        e = NavigationEvent("gethandleside", self.handleside)
-        msg = std_msgs.msg.String()
-        msg.data = str(e)
-        self._eventPub.publish(msg)
-
-    def send_touchmode(self):
-        e = NavigationEvent("gettouchmode", self.touchmode)
-        msg = std_msgs.msg.String()
-        msg.data = str(e)
-        self._eventPub.publish(msg)
-
-    def send_speaker_audio_files(self):
-        if self._interface.audio_files:
-            e = NavigationEvent("getspeakeraudiofiles", self._interface.audio_files)
-            msg = std_msgs.msg.String()
-            msg.data = str(e)
-            self._eventPub.publish(msg)
-        else:
-            return
-
     def create_menu(self):
         try:
             self.create_menu_timer.cancel()
-            menu_file = node.declare_parameter('menu_file', '').value
+            menu_file = self._node.declare_parameter('menu_file', '').value
             with open(menu_file, 'r') as stream:
                 menu_obj = yaml.safe_load(stream)
                 # overwrite init_speed if it is set
@@ -211,8 +184,10 @@ class CabotUIManager(NavigationInterface, object):
     def activity_log(self, category="", text="", memo=""):
         self._interface.activity_log(category, text, memo)
 
+    def change_language(self, lang):
+        self._interface.change_language(lang)
+
     def i_am_ready(self):
-        self._status_manager.set_state(State.idle)
         self._interface.i_am_ready()
 
     def start_navigation(self):
@@ -228,22 +203,6 @@ class CabotUIManager(NavigationInterface, object):
     def notify_human(self, angle=0):
         self._interface.notify_human(angle=angle)
 
-    def goal_canceled(self, goal):
-        # unexpected cancel, may need to retry
-        if self._status_manager.state == State.in_action:
-            self._logger.info("NavigationState: canceled (system)")
-            self._status_manager.set_state(State.in_pausing)
-            self._retry_navigation()
-            return
-        self._logger.info("NavigationState: canceled (user)")
-
-    def _retry_navigation(self):
-        self._retry_count += 1
-        self._logger.info("NavigationState: retrying (system)")
-        self._status_manager.set_state(State.in_action)
-        self._navigation.retry_navigation()
-        self._logger.info("NavigationState: retried (system)")
-
     def have_arrived(self, goal):
         # do not read arrival message from robot
         # self._logger.info("delegate have_arrived called")
@@ -251,7 +210,6 @@ class CabotUIManager(NavigationInterface, object):
         pass
 
     def have_completed(self):
-        self._status_manager.set_state(State.idle)
         # send navigation_arrived event when all goals are completed
         self._logger.info("NavigationState: arrived")
         # notify external nodes about arrival
@@ -314,13 +272,23 @@ class CabotUIManager(NavigationInterface, object):
     def request_sound(self, sound):
         self._interface.request_sound(sound)
 
-    def system_pause_navigation(self, callback):
-        def done_callback():
-            self._status_manager.set_state(State.in_pause)
-            self._logger.info("NavigationState: paused (system)")
-            callback()
-        self._status_manager.set_state(State.in_pausing)
-        self._navigation.pause_navigation(done_callback)
+    def set_pause_control(self, pause: bool):
+        self._interface.set_pause_control(pause)
+
+    def in_preparation(self):
+        self._interface.in_preparation()
+
+    def pause_navigation(self):
+        self._interface.pause_navigation()
+
+    def pausing_navigation(self):
+        self._interface.pausing_navigation()
+
+    def cancel_navigation(self):
+        self._interface.cancel_navigation()
+
+    def resume_navigation(self):
+        self._interface.resume_navigation()
 
     # endregion NavigationInterface
 
@@ -334,25 +302,11 @@ class CabotUIManager(NavigationInterface, object):
         except:  # noqa: #722
             self._logger.error(traceback.format_exc())
 
-    def reset(self):
-        """reset menu"""
-        # if self.main_menu:
-        #     self.main_menu.reset()
-        # self.menu_stack = [self.main_menu]
-
-    # menu delegate method
     def menu_selected(self, menu):
         self._logger.debug(F"menu_selected, {menu.identifier}, {menu.type}")
         if menu.identifier == "destination_menu":
             event = NavigationEvent("destination", menu.value.value)
             self.process_event(event)
-
-        # if menu.identifier == "main_menu" and menu.value is not None:
-        #     self._logger.info(menu.value)
-        #     self._logger.info(menu.value.identifier)
-        #     if menu.value.identifier == "exploration_menu":
-        #         event = ExplorationEvent("start")
-        #         self.process_event(event)
 
     # event delegate method
     def process_event(self, event):
@@ -362,86 +316,15 @@ class CabotUIManager(NavigationInterface, object):
         # self._logger.info(f"process_event {str(event)}")
 
         self._event_mapper.push(event)
-        self._process_menu_event(event)
         self._process_navigation_event(event)
         self._process_exploration_event(event)
-
-    def _process_menu_event(self, event):
-        '''
-        process only menu event
-        '''
-        if event.type != MenuEvent.TYPE:
-            return
-
-        curr_menu = self.menu_stack[-1]
-        if event.subtype == "next":
-            curr_menu.next()
-            self._interface.menu_changed(menu=curr_menu)
-
-        elif event.subtype == "prev":
-            curr_menu.prev()
-            self._interface.menu_changed(menu=curr_menu)
-
-        elif event.subtype == "select":
-            selected = curr_menu.select()
-            if selected is None:  # from main menu
-                if curr_menu.value is None:
-                    curr_menu.next()
-                selected = curr_menu
-            elif not selected.can_explore:
-                self.reset()
-            elif selected is not curr_menu:
-                self.menu_stack.append(selected)
-                if selected.value is None:
-                    selected.next()
-
-            self._interface.menu_changed(menu=selected, usage=True)
-
-        elif event.subtype == "back":
-            if len(self.menu_stack) > 1:
-                self.menu_stack.pop()
-                curr_menu = self.menu_stack[-1]
-
-            self._interface.menu_changed(menu=curr_menu, backed=True)
-
-            self.speed = 0
-            # self.cancel_pub.publish(True)
-            self._navigation.pause_navigation()
 
     def _process_navigation_event(self, event):
         if event.type != NavigationEvent.TYPE:
             return
         self._logger.info(f"_process_navigation_event {event}")
 
-        # operations indepent from the navigation state
-        if event.subtype == "language":
-            self._interface.change_language(event.param)
-            return
-
-        if event.subtype == "reqfeatures":
-            self.send_handleside()
-            self.send_touchmode()
-            self.send_speaker_audio_files()
-            return
-
-        if event.subtype == "handleside":
-            self._logger.info("calling set_handle_side")
-            self._navigation.set_handle_side(event.param)
-            return
-
-        if event.subtype == "touchmode":
-            self._logger.info("calling set_touch_mode")
-            self._navigation.set_touch_mode(event.param)
-            return
-
-        # ignore get event
-        if event.subtype == "getlanguage":
-            return
-        if event.subtype == "gethandleside":
-            return
-        if event.subtype == "gettouchmode":
-            return
-        if event.subtype == "getspeakeraudiofiles":
+        if self._navigation_plugins.process_event(event):
             return
 
         if event.subtype == "speedup":
@@ -460,117 +343,6 @@ class CabotUIManager(NavigationInterface, object):
             msg.data = str(e)
             self._eventPub.publish(msg)
 
-        if event.subtype == "event":
-            self._navigation.process_event(event)
-
-        if event.subtype == "arrived":
-            self.destination = None
-
-        if event.subtype == "decision":
-            if self.destination is None:
-                self._logger.info("NavigationState: Subtour")
-                e = NavigationEvent("subtour", None)
-                msg = std_msgs.msg.String()
-                msg.data = str(e)
-                self._eventPub.publish(msg)
-
-        # deactivate control
-        if event.subtype == "idle":
-            self._logger.info("NavigationState: Pause control = True")
-            self._interface.set_pause_control(True)
-            self._navigation.set_pause_control(True)
-
-        def description_callback(result):
-            try:
-                if result:
-                    self._logger.info(f"description - {result=}")
-                    self._interface.describe_surround(result['translated'])
-                else:
-                    self._logger.info("description - Error")
-                    self._interface.describe_error()
-            except:   # noqa: #722
-                self._logger.error(traceback.format_exc())
-
-        if event.subtype == "description" and self._description.enabled:
-            # TODO: needs to reset last_plan_distance when arrived/paused
-            self._logger.info(F"Request Description duration={event.param}")
-            if self._interface.last_pose:
-                gp = self._interface.last_pose['global_position']
-                cf = self._interface.last_pose['current_floor']
-                length_index = min(2, int(event.param) - 1)   # 1 sec -> 0, 2 sec -> 1, < 3 sec -> 2
-                if self._description.stop_reason_enabled and self._description.surround_enabled:
-                    if length_index <= 1:
-                        self._interface.requesting_describe_surround_stop_reason()
-                    else:
-                        self._interface.requesting_describe_surround()
-                elif self._description.stop_reason_enabled and not self._description.surround_enabled:
-                    self._interface.requesting_describe_surround_stop_reason()
-                elif not self._description.stop_reason_enabled and self._description.surround_enabled:
-                    self._interface.requesting_describe_surround()
-                self._description.request_description_with_images1(gp, cf, self._interface.lang, length_index=length_index, callback=description_callback)
-
-        # request description internal functions
-        def request_stop_reason_description():
-            self._logger.info("description - Request Stop Reason Description")
-            try:
-                if self._interface.last_pose:
-                    gp = self._interface.last_pose['global_position']
-                    cf = self._interface.last_pose['current_floor']
-                    if self._description.request_description_with_images2(
-                            gp,
-                            cf,
-                            "stop_reason",
-                            self._interface.lang,
-                            length_index=0,
-                            callback=description_callback):
-                        self._interface.requesting_describe_surround_stop_reason()
-                    else:
-                        self._interface.requesting_please_wait()
-            except:  # noqa: #722
-                self._logger.error(traceback.format_exc())
-
-        def request_surround_description():
-            self._logger.info(F"description - Request Surround Description (Duration: {event.param})")
-            try:
-                if self._interface.last_pose:
-                    length_index = event.param
-                    gp = self._interface.last_pose['global_position']
-                    cf = self._interface.last_pose['current_floor']
-                    if self._description.request_description_with_images2(
-                            gp,
-                            cf,
-                            "surround",
-                            self._interface.lang,
-                            length_index=length_index,
-                            callback=description_callback):
-                        self._interface.requesting_describe_surround()
-                    else:
-                        self._interface.requesting_please_wait()
-            except:  # noqa: #722
-                self._logger.error(traceback.format_exc())
-
-        if event.subtype == "description_stop_reason" and self._description.enabled:
-            request_stop_reason_description()
-
-        if event.subtype == "description_surround" and self._description.enabled:
-            request_surround_description()
-
-        if event.subtype == "speaker_enable":
-            self._interface.enable_speaker(event.param)
-            return
-
-        if event.subtype == "speaker_alert":
-            self._interface.speaker_alert()
-            return
-
-        if event.subtype == "speaker_audio_file":
-            self._interface.set_audio_file(event.param)
-            return
-
-        if event.subtype == "speaker_volume":
-            self._interface.set_speaker_volume(event.param)
-            return
-
         if event.subtype == "toggle_speak_state":
             self._logger.info("Request Toggle TTS State")
             e = NavigationEvent("togglespeakstate", None)
@@ -585,142 +357,6 @@ class CabotUIManager(NavigationInterface, object):
             msg.data = str(e)
             self._eventPub.publish(msg)
 
-        # operations depents on the current navigation state
-        if self._status_manager.state == State.in_preparation:
-            self.activity_log("cabot_ui/navigation", "in preparation")
-            self._interface.in_preparation()
-            return
-
-        if event.subtype == "destination":
-            if self._status_manager.state != State.idle:
-                self.activity_log("cabot_ui/navigation", "destination", "need to cancel")
-
-                def done_callback():
-                    self._status_manager.set_state(State.idle)
-                    self._process_navigation_event(event)
-                self._navigation.cancel_navigation(done_callback)
-                return
-
-            self._logger.info(F"Destination: {event.param}")
-            self._retry_count = 0
-            self.destination = event.param
-            # change handle mode
-            request = std_srvs.srv.SetBool.Request()
-            request.data = True
-            if self._touchModeProxy.wait_for_service(timeout_sec=1):
-                self._touchModeProxy.call_async(request)
-                # response: std_srvs.srv.SetBool.Response = self._touchModeProxy.call(request)
-                # if not response.success:
-                #     self._logger.error("Could not set touch mode to True")
-            else:
-                self._logger.error("Could not find set touch mode service")
-
-            if self._userSpeedEnabledProxy.wait_for_service(timeout_sec=1):
-                self._userSpeedEnabledProxy.call_async(request)
-                # response = self._userSpeedEnabledProxy.call(request)
-                # if not response.success:
-                #     self._logger.info("Could not set user speed enabled to True")
-            else:
-                self._logger.error("Could not find set user speed enabled service")
-
-            # change state
-            # change to waiting_action by using actionlib
-            self._status_manager.set_state(State.in_action)
-            self._navigation.set_destination(event.param)
-
-        if event.subtype == "summons":
-            if self._status_manager.state != State.idle:
-                self.activity_log("cabot_ui/navigation", "summons", "need to cancel")
-
-                def done_callback():
-                    self._status_manager.set_state(State.idle)
-                    self._process_navigation_event(event)
-                self._navigation.cancel_navigation(done_callback)
-                return
-
-            self._logger.info(F"Summons Destination: {event.param}")
-            self.destination = event.param
-            # change handle mode
-            request = std_srvs.srv.SetBool.Request()
-            request.data = False
-            if self._touchModeProxy.wait_for_service(timeout_sec=1):
-                response: std_srvs.srv.SetBool.Response = self._touchModeProxy.call(request)
-                if not response.success:
-                    self._logger.info("Could not set touch mode to False")
-            else:
-                self._logger.error("Could not find set touch mode service")
-
-            if self._userSpeedEnabledProxy.wait_for_service(timeout_sec=1):
-                response = self._userSpeedEnabledProxy.call(request)
-                if not response.success:
-                    self._logger.info("Could not set user speed enabled to False")
-            else:
-                self._logger.error("Could not find set user speed enabled service")
-
-            # change state
-            # change to waiting_action by using actionlib
-            self._status_manager.set_state(State.in_summons)
-            self._navigation.set_destination(event.param)
-
-        if event.subtype == "cancel":
-            self._logger.info("NavigationState: User Cancel requested")
-            if self._status_manager.state != State.idle:
-                self._logger.info("NavigationState: canceling (user)")
-                self._interface.cancel_navigation()
-
-                def done_callback():
-                    self.in_navigation = False
-                    self.destination = None
-                    self._status_manager.set_state(State.idle)
-                    self._logger.info("NavigationState: canceled (user)")
-                self._navigation.cancel_navigation(done_callback)
-            else:
-                self._logger.info("NavigationState: state is not in action state={}".format(self._status_manager.state))
-
-        if event.subtype == "pause":
-            self._logger.info("NavigationState: User Pause requested")
-            if self._status_manager.state == State.in_action or \
-               self._status_manager.state == State.in_summons:
-                self._logger.info("NavigationState: pausing (user)")
-
-                def done_callback():
-                    self._status_manager.set_state(State.in_pause)
-                    self._logger.info("NavigationState: paused (user)")
-                self._status_manager.set_state(State.in_pausing)
-                self._interface.pause_navigation()
-                self._navigation.pause_navigation(done_callback)
-            else:
-                # force to pause state
-                self._logger.info("NavigationState: state is not in action state={}".format(self._status_manager.state))
-
-        if event.subtype in {"resume", "resume_or_stop_reason"}:
-            if self.destination is not None:
-                self._logger.info("NavigationState: User Resume requested")
-                if self._status_manager.state == State.in_pause:
-                    self._logger.info("NavigationState: resuming (user)")
-                    self._interface.resume_navigation()
-                    self._status_manager.set_state(State.in_action)
-                    self._navigation.resume_navigation()
-                    self._logger.info("NavigationState: resumed (user)")
-                elif self._status_manager.state == State.in_pausing:
-                    self._interface.pausing_navigation()
-                else:
-                    if event.subtype == "resume_or_stop_reason" and self._description.enabled:
-                        request_stop_reason_description()
-                    else:
-                        self._logger.info("NavigationState: state is not in pause state")
-            else:
-                self._logger.info("NavigationState: Next")
-                e = NavigationEvent("next", None)
-                msg = std_msgs.msg.String()
-                msg.data = str(e)
-                self._eventPub.publish(msg)
-
-            # activate control
-            self._logger.info("NavigationState: Pause control = False")
-            self._interface.set_pause_control(False)
-            self._navigation.set_pause_control(False)
-
     def _process_exploration_event(self, event):
         if event.type != ExplorationEvent.TYPE:
             return
@@ -732,47 +368,17 @@ class CabotUIManager(NavigationInterface, object):
 
 class EventMapper1(object):
     def __init__(self):
-        self._manager = StatusManager.get_instance()
         self.description_duration = 0
 
     def push(self, event):
-        # state = self._manager.state
-
         if event.type != ButtonEvent.TYPE and event.type != ClickEvent.TYPE and \
            event.type != HoldDownEvent.TYPE:
             return
 
-        mevent = None
-
         # simplify the control
         mevent = self.map_button_to_navigation(event)
-
-        '''
-        if state == State.idle:
-            mevent = self.map_button_to_menu(event)
-
-        elif state == State.in_action or state == State.waiting_action:
-            mevent = self.map_button_to_navigation(event)
-
-        elif state == State.in_pause or state == State.waiting_pause:
-            mevent = self.map_button_to_navigation(event)
-        '''
-
         if mevent:
             self.delegate.process_event(mevent)
-
-    def map_button_to_menu(self, event):
-        if event.type == "click" and event.count == 1:
-            if event.buttons == cabot_common.button.BUTTON_NEXT:
-                return MenuEvent(subtype="next")
-            if event.buttons == cabot_common.button.BUTTON_PREV:
-                return MenuEvent(subtype="prev")
-            if event.buttons == cabot_common.button.BUTTON_SELECT:
-                return MenuEvent(subtype="select")
-        elif event.type == "click" and event.count == 2:
-            if event.buttons == cabot_common.button.BUTTON_SELECT:
-                return MenuEvent(subtype="back")
-        return None
 
     def map_button_to_navigation(self, event):
         if event.type == "button" and not event.down and self.description_duration > 0:
@@ -798,21 +404,11 @@ class EventMapper1(object):
             if event.holddown == cabot_common.button.BUTTON_RIGHT:
                 # image description is not triggered here, but when button is released
                 self.description_duration = event.duration
-        '''
-        if event.button == cabot_common.button.BUTTON_SELECT:
-                return NavigationEvent(subtype="pause")
-        if event.type == "click":
-            if event.buttons == cabot_common.button.BUTTON_SELECT and event.count == 2:
-                return NavigationEvent(subtype="cancel")
-            if event.buttons == cabot_common.button.BUTTON_NEXT and event.count == 2:
-                return NavigationEvent(subtype="resume")
-        '''
         return None
 
 
 class EventMapper2(object):
     def __init__(self):
-        self._manager = StatusManager.get_instance()
         self.button_hold_down_duration = 0
         self.button_hold_down_duration_prev = 0
 
@@ -820,22 +416,8 @@ class EventMapper2(object):
         if event.type not in {ButtonEvent.TYPE, ClickEvent.TYPE, HoldDownEvent.TYPE}:
             return
 
-        mevent = None
-
         # simplify the control
         mevent = self.map_button_to_navigation(event)
-
-        '''
-        if state == State.idle:
-            mevent = self.map_button_to_menu(event)
-
-        elif state == State.in_action or state == State.waiting_action:
-            mevent = self.map_button_to_navigation(event)
-
-        elif state == State.in_pause or state == State.waiting_pause:
-            mevent = self.map_button_to_navigation(event)
-        '''
-
         if mevent:
             self.delegate.process_event(mevent)
 
@@ -889,9 +471,8 @@ class EventMapper2(object):
 
 def receiveSignal(signal_num, frame):
     print("Received:", signal_num)
-    node.destroy_node()
-    for t in threads:
-        t.join()
+    # node.destroy_node()
+    node_manager.join()
     sys.exit()
 
 
@@ -900,58 +481,9 @@ signal.signal(signal.SIGINT, receiveSignal)
 
 if __name__ == "__main__":
     rclpy.init()
-    node = Node('cabot_ui_manager', start_parameter_services=False)
-    nav_node = Node("cabot_ui_manager_navigation", start_parameter_services=False)
-    tf_node = Node("cabot_ui_manager_tf", start_parameter_services=False)
-    srv_node = Node("cabot_ui_manager_navigation_service", start_parameter_services=False)
-    act_node = Node("cabot_ui_manager_navigation_actions", start_parameter_services=False)
-    soc_node = Node("cabot_ui_manager_navigation_social", start_parameter_services=False)
-    desc_node = Node("cabot_ui_manager_description", start_parameter_services=False)
-    nodes = [node, nav_node, tf_node, srv_node, act_node, soc_node, desc_node]
-    executors = [MultiThreadedExecutor(),
-                 MultiThreadedExecutor(),
-                 SingleThreadedExecutor(),
-                 SingleThreadedExecutor(),
-                 SingleThreadedExecutor(),
-                 SingleThreadedExecutor(),
-                 SingleThreadedExecutor(),
-                 ]
-    names = ["node", "tf", "nav", "srv", "act", "soc", "desc"]
-    manager = CabotUIManager(node, nav_node, tf_node, srv_node, act_node, soc_node, desc_node)
 
-    threads = []
-    for tnode, executor, name in zip(nodes, executors, names):
-        def run_node(target_node, executor, name):
-            def _run_node():
-                # debug code to analyze the bottle neck of nodes
-                # high frequency spinning node should have smaller number of waits
-                #
-                # import time
-                # count = 0
-                # start = time.time()
-                executor.add_node(target_node)
-                try:
-                    while rclpy.ok():
-                        # count += 1
-                        # target_node.get_logger().info(f"spin rate {name} {count / (time.time()-start):.2f}Hz - \n"
-                        #                               f"  subscriptions {[sub.topic_name for sub in list(target_node.subscriptions)]}\n"
-                        #                               f"  timers {len(list(target_node.timers))}\n"
-                        #                               f"  clients {[cli.srv_name for cli in list(target_node.clients)]}\n"
-                        #                               f"  services {len(list(target_node.services))}\n"
-                        #                               f"  guards {len(list(target_node.guards))}\n"
-                        #                               f"  waitables {len(list(target_node.waitables))}\n",
-                        #                               throttle_duration_sec=1.0)
-                        executor.spin_once()
-                except KeyboardInterrupt:
-                    target_node.get_logger().info(f"Shutting down {name} node")
-                except:  # noqa: 722
-                    target_node.get_logger().error(traceback.format_exc())
-                target_node.destroy_node()
-            return _run_node
+    node_manager = NodeManager()
+    ui_manager = CabotUIManager(node_manager)
+    node_manager.join()
+    node_manager.get_node().get_logger().info("cabot_ui_manager terminated")
 
-        thread = threading.Thread(target=run_node(tnode, executor, name))
-        thread.start()
-        threads.append(thread)
-
-    for thread in threads:
-        thread.join()
