+++ cabot_ui_plugins/cabot_ui_plugins/description/__init__.py
@@ -1,0 +1,3 @@
+from .description import Description
+
+__all__ = ['Description']
