+++ cabot_ui_plugins/cabot_ui_plugins/feature/__init__.py
@@ -1,0 +1,4 @@
+from .feature import Feature
+
+
+__all__ = ['Feature']
