+++ cabot_ui_plugins/cabot_ui_plugins/navigation/__init__.py
@@ -1,0 +1,9 @@
+from .navgoal import NavGoal
+from .navgoal import make_goals
+from .navigation import Navigation
+
+__all__ = [
+    'NavGoal',
+    'make_goals'
+    'Navigation',
+]
