+++ cabot_ui_plugins/cabot_ui_plugins/speaker/__init__.py
@@ -1,0 +1,3 @@
+from .speaker import Speaker
+
+__all__ = ['Speaker']
