+++ docker-compose-common.yaml
@@ -286,6 +286,7 @@ services:
       - ./cabot_navigation2:/home/developer/ros2_ws/src/cabot_navigation2
       - ./cabot_sites:/home/developer/ros2_ws/src/cabot_sites
       - ./cabot_ui:/home/developer/ros2_ws/src/cabot_ui
+      - ./cabot_ui_plugins:/home/developer/ros2_ws/src/cabot_ui_plugins
       - ./mf_localization_msgs:/home/developer/ros2_ws/src/mf_localization_msgs
       - ./mf_localization_rviz:/home/developer/ros2_ws/src/mf_localization_rviz
       - ./pedestrian_plugin:/home/developer/ros2_ws/src/pedestrian_plugin
