+++ docker/ros2/Dockerfile
@@ -192,6 +192,7 @@ COPY --chown=$USERNAME:$USERNAME --from=cabot_src ./cabot_diagnostics src/cabot_
 COPY --chown=$USERNAME:$USERNAME --from=cabot_src ./cabot_gazebo src/cabot_gazebo
 COPY --chown=$USERNAME:$USERNAME --from=cabot_src ./cabot_navigation2 src/cabot_navigation2
 COPY --chown=$USERNAME:$USERNAME --from=cabot_src ./cabot_ui src/cabot_ui
+COPY --chown=$USERNAME:$USERNAME --from=cabot_src ./cabot_ui_plugins src/cabot_ui_plugins
 COPY --chown=$USERNAME:$USERNAME --from=cabot_src ./mf_localization_msgs src/mf_localization_msgs
 COPY --chown=$USERNAME:$USERNAME --from=cabot_src ./mf_localization_rviz src/mf_localization_rviz
 COPY --chown=$USERNAME:$USERNAME --from=cabot_src ./pedestrian_plugin src/pedestrian_plugin
